Add files via upload
This R script includes interquartile range (IQR) and K-Nearest-Neighbors methods for the detection of outliers for several variables. To that end, I employ several functions and for-loops to make the code simpler, The repetition of the process for IQR outlier detection for the same variables in several subsets of the original dataset makes the use of functions and for-loops useful for the readability of the script.